chore: rename files to snake_case
This is the naming convention recommended for python in PEP8
and most often used for shell scripts.
Helps consistency, readability, in some cases makes the name
easier to type.
Log files have also been renamed accordingly.
