+++ bin/extractors/dratf_extractor.py
