+++ bin/message_factory/message_factory.py
