+++ bin/run_generate_main_page.sh
