+++ bin/utility/static_variables.py
