+++ bin/yang_exclude_bad_drafts.py
